+++ src/main/scala/org/mmisw/orr/ont/app/BaseOntController.scala
@@ -85,6 +85,23 @@ with Logging {
   protected def getFormatParameter: Option[String] =
     params.get("format") orElse params.get("_format")
 
+  protected def getRequestedFormat: Option[String] = {
+    def getAcceptHeader = {
+      val ah = acceptHeader
+      if (logger.underlying.isDebugEnabled) {
+        logger.debug(s"raw Accept: ${request.headers.get("Accept")}")
+        logger.debug(s"acceptHeader: $ah")
+      }
+      ah
+    }
+    getFormatParameter orElse (getAcceptHeader match {
+      case Nil | List("*/*")                      ⇒ None
+      case list if list contains "text/html"      ⇒ Some("html")
+      case list if mimeTypes.contains(list.head)  ⇒ Some(mimeTypes(list.head))
+      case _ ⇒ None
+    })
+  }
+
   protected def completeOntologyUriResolution(ont: Ontology, reqFormatOpt: Option[String] = None) = {
     val versionOpt: Option[String] = params.get("version")
     val (ontVersion, version) = resolveOntologyVersion(ont, versionOpt)